created abstract class RowBy
