+++ RecordParser/Extensions/FileReader/RowReaders/RowByQuote.cs
@@ -1,60 +1,19 @@
 ﻿using System;
-using System.Buffers;
 using System.Collections.Generic;
 using System.IO;
 
-namespace RecordParser.Extensions.FileReader
+namespace RecordParser.Extensions.FileReader.RowReaders
 {
-    internal class RowByQuote : IFL
+    internal class RowByQuote : RowBy
     {
-        private int i = 0;
-        private int j = 0;
-        private int c;
-        private TextReader reader;
-        private bool initial = true;
-
-        private bool yieldLast = false;
-
-        private int bufferLength;
-        private char[] buffer;
         public readonly string separator;
 
         public RowByQuote(TextReader reader, int bufferLength, string separator)
+            : base(reader, bufferLength)
         {
-            this.reader = reader;
-
-            buffer = ArrayPool<char>.Shared.Rent(bufferLength);
-            this.bufferLength = buffer.Length;
-
             this.separator = separator;
         }
-
-        public int FillBuffer()
-        {
-            var len = i - j;
-            if (initial == false)
-            {
-                Array.Copy(buffer, j, buffer, 0, len);
-            }
-
-            var totalRead = reader.Read(buffer, len, bufferLength - len);
-            bufferLength = len + totalRead;
-
-            i = 0;
-            j = 0;
-
-            initial = false;
-
-            if (totalRead == 0 && len != 0 && yieldLast == false)
-            {
-                yieldLast = true;
-                return len;
-            }
-
-            return totalRead;
-        }
-
-        public IEnumerable<ReadOnlyMemory<char>> ReadLines()
+        public override IEnumerable<ReadOnlyMemory<char>> ReadLines()
         {
             int Peek() => i < bufferLength ? buffer[i] : -1;
 
@@ -70,8 +29,6 @@ namespace RecordParser.Extensions.FileReader
             while (hasBufferToConsume = i < bufferLength)
             {
                 c = buffer[i++];
-
-                //  ReadOnlySpan<char> look = buffer.AsSpan().Slice(j, i - j);
 
                 // '\r' => 13
                 // '\n' => 10
@@ -109,7 +66,7 @@ namespace RecordParser.Extensions.FileReader
                         var next = unlook.Slice(z + 1);
 
                         if (next.IsEmpty)
-                            ; // sdfdsf;
+                            ; 
 
                         if (next[0] == quote)
                         {
@@ -145,15 +102,6 @@ namespace RecordParser.Extensions.FileReader
 
             yield return buffer.AsMemory(j, i - j);
             goto reloop;
-        }
-
-        public void Dispose()
-        {
-            if (buffer != null)
-            {
-                ArrayPool<char>.Shared.Return(buffer);
-                buffer = null;
-            }
         }
     }
 }
